apps/sample-apps/nua-flask-postgresql/pyproject.toml
[tool.poetry]
name = "nua-flask-postgresql"
version = "1.0"
description = "Flask + Postgres demo app"
authors = [
    "Jerome Dumonteil <jd@abilian.com>",
    "Stefane Fermigier <sf@abilian.com>",
]
license = "MIT"
packages = [
    { include = "flask_app", from = "src" },
]

[tool.poetry.scripts]
init-db = "flask_app.init_db:main"

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
Flask = "^2.1.3"
Flask-WTF = "^1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
abilian-devtools = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
